Split GGUF and ModelLoader
