+++ Llama3.java
@@ -168,103 +168,6 @@ public class Llama3 {
         }
     }
 
-    record Options(Path modelPath, String prompt, String systemPrompt, boolean interactive,
-                   float temperature, float topp, long seed, int maxTokens, boolean stream, boolean echo) {
-
-        static final int DEFAULT_MAX_TOKENS = 512;
-
-        Options {
-            require(modelPath != null, "Missing argument: --model <path> is required");
-            require(interactive || prompt != null, "Missing argument: --prompt is required in --instruct mode e.g. --prompt \"Why is the sky blue?\"");
-            require(0 <= temperature, "Invalid argument: --temperature must be non-negative");
-            require(0 <= topp && topp <= 1, "Invalid argument: --top-p must be within [0, 1]");
-        }
-
-        static void require(boolean condition, String messageFormat, Object... args) {
-            if (!condition) {
-                System.out.println("ERROR " + messageFormat.formatted(args));
-                System.out.println();
-                printUsage(System.out);
-                System.exit(-1);
-            }
-        }
-
-        static void printUsage(PrintStream out) {
-            out.println("Usage:  jbang Llama3.java [options]");
-            out.println();
-            out.println("Options:");
-            out.println("  --model, -m <path>            required, path to .gguf file");
-            out.println("  --interactive, --chat, -i     run in chat mode");
-            out.println("  --instruct                    run in instruct (once) mode, default mode");
-            out.println("  --prompt, -p <string>         input prompt");
-            out.println("  --system-prompt, -sp <string> (optional) system prompt");
-            out.println("  --temperature, -temp <float>  temperature in [0,inf], default 0.1");
-            out.println("  --top-p <float>               p value in top-p (nucleus) sampling in [0,1] default 0.95");
-            out.println("  --seed <long>                 random seed, default System.nanoTime()");
-            out.println("  --max-tokens, -n <int>        number of steps to run for < 0 = limited by context length, default " + DEFAULT_MAX_TOKENS);
-            out.println("  --stream <boolean>            print tokens during generation; may cause encoding artifacts for non ASCII text, default true");
-            out.println("  --echo <boolean>              print ALL tokens to stderr, if true, recommended to set --stream=false, default false");
-            out.println();
-            out.println("Examples:");
-            out.println("  jbang Llama3.java --model llama3.2-1b-q4_0.gguf --prompt \"Tell me a joke\"");
-            out.println("  jbang Llama3.java --model llama3.2-1b-q4_0.gguf --system-prompt \"Reply concisely, in French\" --prompt \"Who was Marie Curie?\"");
-            out.println("  jbang Llama3.java --model llama3.2-1b-q4_0.gguf --system-prompt \"Answer concisely\" --chat");
-            out.println("  jbang Llama3.java --model llama3.2-1b-q4_0.gguf --chat");
-            out.println("  jbang Llama3.java --model llama3.2-1b-q4_0.gguf --prompt \"Print 5 emojis\" --stream=false");
-        }
-
-        static Options parseOptions(String[] args) {
-            String prompt = null;
-            String systemPrompt = null;
-            float temperature = 0.1f;
-            float topp = 0.95f;
-            Path modelPath = null;
-            long seed = System.nanoTime();
-            // Keep max context length small for low-memory devices.
-            int maxTokens = DEFAULT_MAX_TOKENS;
-            boolean interactive = false;
-            boolean stream = true;
-            boolean echo = false;
-
-            for (int i = 0; i < args.length; i++) {
-                String optionName = args[i];
-                require(optionName.startsWith("-"), "Invalid option %s", optionName);
-                switch (optionName) {
-                    case "--interactive", "--chat", "-i" -> interactive = true;
-                    case "--instruct" -> interactive = false;
-                    case "--help", "-h" -> {
-                        printUsage(System.out);
-                        System.exit(0);
-                    }
-                    default -> {
-                        String nextArg;
-                        if (optionName.contains("=")) {
-                            String[] parts = optionName.split("=", 2);
-                            optionName = parts[0];
-                            nextArg = parts[1];
-                        } else {
-                            require(i + 1 < args.length, "Missing argument for option %s", optionName);
-                            nextArg = args[i + 1];
-                            i += 1; // skip arg
-                        }
-                        switch (optionName) {
-                            case "--prompt", "-p" -> prompt = nextArg;
-                            case "--system-prompt", "-sp" -> systemPrompt = nextArg;
-                            case "--temperature", "--temp" -> temperature = Float.parseFloat(nextArg);
-                            case "--top-p" -> topp = Float.parseFloat(nextArg);
-                            case "--model", "-m" -> modelPath = Paths.get(nextArg);
-                            case "--seed", "-s" -> seed = Long.parseLong(nextArg);
-                            case "--max-tokens", "-n" -> maxTokens = Integer.parseInt(nextArg);
-                            case "--stream" -> stream = Boolean.parseBoolean(nextArg);
-                            case "--echo" -> echo = Boolean.parseBoolean(nextArg);
-                            default -> require(false, "Unknown option: %s", optionName);
-                        }
-                    }
-                }
-            }
-            return new Options(modelPath, prompt, systemPrompt, interactive, temperature, topp, seed, maxTokens, stream, echo);
-        }
-    }
 
     public static void main(String[] args) throws IOException {
         Options options = Options.parseOptions(args);
@@ -282,539 +185,8 @@ public class Llama3 {
     }
 }
 
-final class GGUF {
-    private static final int GGUF_MAGIC = 0x46554747;
-    private static final int DEFAULT_ALIGNMENT = 32; // must be a power of 2
-    private static final List<Integer> SUPPORTED_GGUF_VERSIONS = List.of(2, 3);
-    private int magic;
-    private int version;
-    private int tensorCount; // uint64_t
-    private int alignment;
-    private int metadata_kv_count; // uint64_t
-    private Map<String, Object> metadata;
 
-    public Map<String, GGUFTensorInfo> getTensorInfos() {
-        return tensorInfos;
-    }
 
-    private Map<String, GGUFTensorInfo> tensorInfos;
-
-    private long tensorDataOffset;
-
-    public long getTensorDataOffset() {
-        return tensorDataOffset;
-    }
-
-    public Map<String, Object> getMetadata() {
-        return metadata;
-    }
-
-    private final ByteBuffer BB_1 = ByteBuffer.allocate(Byte.BYTES).order(ByteOrder.LITTLE_ENDIAN);
-    private final ByteBuffer BB_2 = ByteBuffer.allocate(Short.BYTES).order(ByteOrder.LITTLE_ENDIAN);
-    private final ByteBuffer BB_4 = ByteBuffer.allocate(Integer.BYTES).order(ByteOrder.LITTLE_ENDIAN);
-    private final ByteBuffer BB_8 = ByteBuffer.allocate(Long.BYTES).order(ByteOrder.LITTLE_ENDIAN);
-
-    public static GGUF loadModel(Path modelPath) throws IOException {
-        try (FileChannel fileChannel = FileChannel.open(modelPath);
-             var ignored = Timer.log("Parse " + modelPath)) {
-            GGUF gguf = new GGUF();
-            gguf.loadModelImpl(fileChannel);
-            return gguf;
-        }
-    }
-
-    enum MetadataValueType {
-        // The value is a 8-bit unsigned integer.
-        UINT8(1),
-        // The value is a 8-bit signed integer.
-        INT8(1),
-        // The value is a 16-bit unsigned little-endian integer.
-        UINT16(2),
-        // The value is a 16-bit signed little-endian integer.
-        INT16(2),
-        // The value is a 32-bit unsigned little-endian integer.
-        UINT32(4),
-        // The value is a 32-bit signed little-endian integer.
-        INT32(4),
-        // The value is a 32-bit IEEE754 floating point number.
-        FLOAT32(4),
-        // The value is a boolean.
-        // 1-byte value where 0 is false and 1 is true.
-        // Anything else is invalid, and should be treated as either the model being invalid or the reader being buggy.
-        BOOL(1),
-        // The value is a UTF-8 non-null-terminated string, with length prepended.
-        STRING(-8),
-        // The value is an array of other values, with the length and type prepended.
-        // Arrays can be nested, and the length of the array is the number of elements in the array, not the number of bytes.
-        ARRAY(-8),
-        // The value is a 64-bit unsigned little-endian integer.
-        UINT64(8),
-        // The value is a 64-bit signed little-endian integer.
-        INT64(8),
-        // The value is a 64-bit IEEE754 floating point number.
-        FLOAT64(8);
-        private final int byteSize;
-
-        MetadataValueType(int byteSize) {
-            this.byteSize = byteSize;
-        }
-
-        private static final MetadataValueType[] VALUES = values();
-
-        public static MetadataValueType fromIndex(int index) {
-            return VALUES[index];
-        }
-
-        public int byteSize() {
-            return byteSize;
-        }
-    }
-
-    private void loadModelImpl(FileChannel fileChannel) throws IOException {
-        // The header of the file.
-        readHeader(fileChannel); // gguf_header_t header;
-        // Tensor infos, which can be used to locate the tensor data.
-        // gguf_tensor_info_t tensor_infos[header.tensor_count];
-        this.tensorInfos = HashMap.newHashMap(tensorCount);
-        for (int i = 0; i < tensorCount; ++i) {
-            GGUF.GGUFTensorInfo ti = readTensorInfo(fileChannel);
-            assert !tensorInfos.containsKey(ti.name);
-            tensorInfos.put(ti.name, ti);
-        }
-        // Padding to the nearest multiple of `ALIGNMENT`.
-        // uint8_t _padding[ALIGNMENT - (sizeof(header + tensor_infos) % ALIGNMENT)];
-        //long _padding = -fileChannel.position() & (ALIGNMENT - 1);
-        long _padding = getAlignment() - (fileChannel.position() % getAlignment());
-        fileChannel.position(fileChannel.position() + _padding);
-        // Tensor data.
-        //
-        // This is arbitrary binary data corresponding to the weights of the model. This data should be close
-        // or identical to the data in the original model file, but may be different due to quantization or
-        // other optimizations for inference. Any such deviations should be recorded in the metadata or as
-        // part of the architecture definition.
-        //
-        // Each tensor's data must be stored within this array, and located through its `tensor_infos` entry.
-        // The offset of each tensor's data must be a multiple of `ALIGNMENT`, and the space between tensors
-        // should be padded to `ALIGNMENT` bytes.
-        // uint8_t tensor_data[];
-        this.tensorDataOffset = fileChannel.position();
-    }
-
-    public static Map<String, GGMLTensorEntry> loadTensors(FileChannel fileChannel, long tensorDataOffset, Map<String, GGUFTensorInfo> tensorInfos) throws IOException {
-        Arena arena = Arena.ofAuto();
-        MemorySegment tensorData = fileChannel.map(FileChannel.MapMode.READ_ONLY, tensorDataOffset, fileChannel.size() - tensorDataOffset, arena);
-        Map<String, GGMLTensorEntry> tensorEntries = HashMap.newHashMap(tensorInfos.size());
-        for (Map.Entry<String, GGUFTensorInfo> entry : tensorInfos.entrySet()) {
-            GGUFTensorInfo ti = entry.getValue();
-            int numberOfElements = FloatTensor.numberOfElements(ti.dimensions());
-            int sizeInBytes = Math.toIntExact(ti.ggmlType().byteSizeFor(numberOfElements));
-            MemorySegment memorySegment = tensorData.asSlice(ti.offset(), sizeInBytes);
-            tensorEntries.put(ti.name(), new GGMLTensorEntry(tensorData, ti.name(), ti.ggmlType(), ti.dimensions(), memorySegment));
-        }
-        return tensorEntries;
-    }
-
-    public record GGUFTensorInfo(String name, int[] dimensions, GGMLType ggmlType, long offset) {
-    }
-
-    private GGMLType readGGMLType(FileChannel fileChannel) throws IOException {
-        int ggmlTypeId = readInt(fileChannel); // ggml_type type;
-        return GGMLType.fromId(ggmlTypeId);
-    }
-
-    private GGUF.GGUFTensorInfo readTensorInfo(FileChannel fileChannel) throws IOException {
-        // The name of the tensor. It is a standard GGUF string, with the caveat that
-        // it must be at most 64 bytes long.
-        String name = readString(fileChannel); // gguf_string_t name;
-        assert name.length() <= 64;
-        // The number of dimensions in the tensor.
-        // Currently at most 4, but this may change in the future.
-        int n_dimensions = readInt(fileChannel); // uint32_t n_dimensions;
-        assert n_dimensions <= 4;
-        // The dimensions of the tensor.
-        int[] dimensions = new int[n_dimensions]; // uint64_t dimensions[n_dimensions];
-        for (int i = 0; i < n_dimensions; ++i) {
-            dimensions[i] = Math.toIntExact(readLong(fileChannel));
-        }
-        // The type of the tensor.
-        GGMLType ggmlType = readGGMLType(fileChannel); // ggml_type type;
-        // The offset of the tensor's data in this file in bytes.
-        // This offset is relative to `tensor_data`, not to the start
-        // of the file, to make it easier for writers to write the file.
-        // Readers should consider exposing this offset relative to the
-        // file to make it easier to read the data.
-        // Must be a multiple of `ALIGNMENT`.
-        long offset = readLong(fileChannel); // uint64_t offset;
-        assert offset % getAlignment() == 0;
-        return new GGUF.GGUFTensorInfo(name, dimensions, ggmlType, offset);
-    }
-
-    private String readString(FileChannel fileChannel) throws IOException {
-        // A string in GGUF.
-        // The length of the string, in bytes.
-        int len = Math.toIntExact(readLong(fileChannel)); // uint64_t len;
-        // The string as a UTF-8 non-null-terminated string.
-        byte[] bytes = new byte[len]; // char string[len];
-        int bytesRead = fileChannel.read(ByteBuffer.wrap(bytes));
-        assert len == bytesRead;
-        return new String(bytes, StandardCharsets.UTF_8);
-    }
-
-    private Pair<String, Object> readKeyValuePair(FileChannel fileChannel) throws IOException {
-        // The key of the metadata. It is a standard GGUF string, with the following caveats:
-        // - It must be a valid ASCII string.
-        // - It must be a hierarchical key, where each segment is `lower_snake_case` and separated by a `.`.
-        // - It must be at most 2^16-1/65535 bytes long.
-        // Any keys that do not follow these rules are invalid.
-        String key = readString(fileChannel); // gguf_string_t key;
-        assert key.length() < (1 << 16);
-        assert key.codePoints().allMatch(cp -> ('a' <= cp && cp <= 'z') || ('0' <= cp && cp <= '9') || cp == '_' || cp == '.');
-        Object value = readMetadataValue(fileChannel);
-        return new Pair<>(key, value);
-    }
-
-    private Object readMetadataValue(FileChannel fileChannel) throws IOException {
-        // The type of the value.
-        // Must be one of the `gguf_metadata_value_type` values.
-        MetadataValueType value_type = readMetadataValueType(fileChannel); // gguf_metadata_value_type value_type;
-        // The value.
-        return readMetadataValueOfType(value_type, fileChannel); // gguf_metadata_value_t value;
-    }
-
-    void readHeader(FileChannel fileChannel) throws IOException {
-        // Magic number to announce that this is a GGUF file.
-        // Must be `GGUF` at the byte level: `0x47` `0x47` `0x55` `0x46`.
-        // Your executor might do little-endian byte order, so it might be
-        // check for 0x46554747 and letting the endianness cancel out.
-        // Consider being *very* explicit about the byte order here.
-        this.magic = readInt(fileChannel); //    uint32_t magic;
-        if (magic != GGUF_MAGIC) {
-            throw new IllegalArgumentException("unsupported header.magic " + magic);
-        }
-        // The version of the format implemented.
-        // Must be `3` for version described in this spec.
-        //
-        // This version should only be increased for structural changes to the format.
-        // Changes that do not affect the structure of the file should instead update the metadata
-        // to signify the change.
-        this.version = readInt(fileChannel); // uint32_t version;
-        if (!SUPPORTED_GGUF_VERSIONS.contains(version)) {
-            throw new IllegalArgumentException("unsupported header.version " + version);
-        }
-        // The number of tensors in the file.
-        // This is explicit, instead of being included in the metadata, to ensure it is always present
-        // for loading the tensors.
-        this.tensorCount = Math.toIntExact(readLong(fileChannel)); // uint64_t tensor_count;
-        // The number of metadata key-value pairs.
-        this.metadata_kv_count = Math.toIntExact(readLong(fileChannel)); // uint64_t metadata_kv_count;
-        // The metadata key-value pairs.
-        // gguf_metadata_kv_t metadata_kv[metadata_kv_count];
-        this.metadata = HashMap.newHashMap(metadata_kv_count);
-        for (int i = 0; i < metadata_kv_count; ++i) {
-            Pair<String, Object> keyValue = readKeyValuePair(fileChannel);
-            assert !metadata.containsKey(keyValue.first());
-            metadata.put(keyValue.first(), keyValue.second());
-        }
-    }
-
-    private Object readArray(FileChannel fileChannel) throws IOException {
-        // Any value type is valid, including arrays.
-        MetadataValueType value_type = readMetadataValueType(fileChannel); // gguf_metadata_value_type type;
-        // Number of elements, not bytes
-        int len = Math.toIntExact(readLong(fileChannel)); // uint64_t len;
-        // The array of values.
-        // gguf_metadata_value_t array[len];
-        switch (value_type) {
-            case UINT8, INT8 -> {
-                byte[] bytes = new byte[len];
-                for (int i = 0; i < len; ++i) {
-                    bytes[i] = readByte(fileChannel);
-                }
-                return bytes;
-            }
-            case UINT16, INT16 -> {
-                short[] shorts = new short[len];
-                for (int i = 0; i < len; ++i) {
-                    shorts[i] = readShort(fileChannel);
-                }
-                return shorts;
-            }
-            case UINT32, INT32 -> {
-                int[] ints = new int[len];
-                for (int i = 0; i < len; ++i) {
-                    ints[i] = readInt(fileChannel);
-                }
-                return ints;
-            }
-            case FLOAT32 -> {
-                float[] floats = new float[len];
-                for (int i = 0; i < len; ++i) {
-                    floats[i] = readFloat(fileChannel);
-                }
-                return floats;
-            }
-            case BOOL -> {
-                boolean[] booleans = new boolean[len];
-                for (int i = 0; i < len; ++i) {
-                    booleans[i] = readBoolean(fileChannel);
-                }
-                return booleans;
-            }
-            case STRING -> {
-                String[] strings = new String[len];
-                for (int i = 0; i < len; ++i) {
-                    strings[i] = readString(fileChannel);
-                }
-                return strings;
-            }
-            case ARRAY -> {
-                Object[] arrays = new Object[len];
-                for (int i = 0; i < len; ++i) {
-                    arrays[i] = readArray(fileChannel);
-                }
-                return arrays;
-            }
-            default -> throw new UnsupportedOperationException("read array of " + value_type);
-        }
-    }
-
-    private Object readMetadataValueOfType(MetadataValueType valueType, FileChannel fileChannel) throws IOException {
-        return switch (valueType) {
-            case UINT8, INT8 -> readByte(fileChannel);
-            case UINT16, INT16 -> readShort(fileChannel);
-            case UINT32, INT32 -> readInt(fileChannel);
-            case FLOAT32 -> readFloat(fileChannel);
-            case UINT64, INT64 -> readLong(fileChannel);
-            case FLOAT64 -> readDouble(fileChannel);
-            case BOOL -> readBoolean(fileChannel);
-            case STRING -> readString(fileChannel);
-            case ARRAY -> readArray(fileChannel);
-        };
-    }
-
-    private byte readByte(FileChannel fileChannel) throws IOException {
-        int bytesRead = fileChannel.read(BB_1);
-        assert bytesRead == 1;
-        return BB_1.clear().get(0);
-    }
-
-    private boolean readBoolean(FileChannel fileChannel) throws IOException {
-        return readByte(fileChannel) != 0;
-    }
-
-    private short readShort(FileChannel fileChannel) throws IOException {
-        int bytesRead = fileChannel.read(BB_2);
-        assert bytesRead == 2;
-        return BB_2.clear().getShort(0);
-    }
-
-    private int readInt(FileChannel fileChannel) throws IOException {
-        int bytesRead = fileChannel.read(BB_4);
-        assert bytesRead == 4;
-        return BB_4.clear().getInt(0);
-    }
-
-    private long readLong(FileChannel fileChannel) throws IOException {
-        int bytesRead = fileChannel.read(BB_8);
-        assert bytesRead == 8;
-        return BB_8.clear().getLong(0);
-    }
-
-    private float readFloat(FileChannel fileChannel) throws IOException {
-        return Float.intBitsToFloat(readInt(fileChannel));
-    }
-
-    private double readDouble(FileChannel fileChannel) throws IOException {
-        return Double.longBitsToDouble(readLong(fileChannel));
-    }
-
-    private MetadataValueType readMetadataValueType(FileChannel fileChannel) throws IOException {
-        int index = readInt(fileChannel);
-        return MetadataValueType.fromIndex(index);
-    }
-
-    public int getAlignment() {
-        if (alignment != 0) {
-            return alignment;
-        }
-        alignment = (int) metadata.getOrDefault("general.alignment", DEFAULT_ALIGNMENT);
-        assert Integer.bitCount(alignment) == 1 : "alignment must be a power of two";
-        return alignment;
-    }
-}
-
-interface Timer extends AutoCloseable {
-    @Override
-    void close(); // no Exception
-
-    static Timer log(String label) {
-        return log(label, TimeUnit.MILLISECONDS);
-    }
-
-    static Timer log(String label, TimeUnit timeUnit) {
-        return new Timer() {
-            final long startNanos = System.nanoTime();
-
-            @Override
-            public void close() {
-                long elapsedNanos = System.nanoTime() - startNanos;
-                System.err.println(label + ": "
-                        + timeUnit.convert(elapsedNanos, TimeUnit.NANOSECONDS) + " "
-                        + timeUnit.toChronoUnit().name().toLowerCase());
-            }
-        };
-    }
-}
-
-final class ModelLoader {
-    private static final String TOKENIZER_LLAMA_3_MODEL = "gpt2";
-
-    private static final String LLAMA_3_PATTERN = "(?i:'s|'t|'re|'ve|'m|'ll|'d)|[^\\r\\n\\p{L}\\p{N}]?\\p{L}+|\\p{N}{1,3}| ?[^\\s\\p{L}\\p{N}]+[\\r\\n]*|\\s*[\\r\\n]+|\\s+(?!\\S)|\\s+";
-
-    private static Vocabulary loadVocabulary(Map<String, Object> metadata) {
-        String model = (String) metadata.get("tokenizer.ggml.model");
-        if (!TOKENIZER_LLAMA_3_MODEL.equals(model)) {
-            throw new IllegalArgumentException("expected " + TOKENIZER_LLAMA_3_MODEL + " but found " + model);
-        }
-        String[] tokens = (String[]) metadata.get("tokenizer.ggml.tokens");
-        return new Vocabulary(tokens, null);
-    }
-
-    public static Llama loadModel(Path ggufPath, int contextLength, boolean loadWeights) throws IOException {
-        GGUF gguf = GGUF.loadModel(ggufPath);
-        FileChannel fileChannel = FileChannel.open(ggufPath, StandardOpenOption.READ);
-        return loadModel(fileChannel, gguf, contextLength, loadWeights);
-    }
-
-    public static Llama loadModel(FileChannel fileChannel, GGUF gguf, int contextLength, boolean loadWeights) throws IOException {
-        try (var ignored = Timer.log("Load LlaMa model")) {
-            Map<String, Object> metadata = gguf.getMetadata();
-            Vocabulary vocabulary = loadVocabulary(metadata);
-            Tokenizer tokenizer = createTokenizer(metadata, vocabulary);
-
-            Llama.Configuration config = new Llama.Configuration(
-                    (int) metadata.get("llama.embedding_length"),
-                    (int) metadata.get("llama.feed_forward_length"),
-                    (int) metadata.get("llama.block_count"),
-                    (int) metadata.get("llama.attention.head_count"),
-
-                    metadata.containsKey("llama.attention.head_count_kv")
-                            ? (int) metadata.get("llama.attention.head_count_kv")
-                            : (int) metadata.get("llama.attention.head_count"),
-
-                    vocabulary.size(),
-                    (int) metadata.get("llama.context_length"),
-                    (float) metadata.getOrDefault("llama.attention.layer_norm_rms_epsilon", 1e-5f),
-                    (float) metadata.getOrDefault("llama.rope.freq_base", 10000f)
-            ).withContextLength(contextLength);
-
-            Llama.Weights weights = null;
-            if (loadWeights) {
-                Map<String, GGMLTensorEntry> tensorEntries = GGUF.loadTensors(fileChannel, gguf.getTensorDataOffset(), gguf.getTensorInfos());
-                weights = loadWeights(tensorEntries, config);
-            }
-            return new Llama(config, tokenizer, weights);
-        }
-    }
-
-    static Llama.Weights loadWeights(Map<String, GGMLTensorEntry> tensorEntries, Llama.Configuration config) {
-        boolean ropeScaling = tensorEntries.containsKey("rope_freqs");
-        float scaleFactor = 8;
-        float loFreqFactor = 1;
-        float hiFreqFactor = 3;
-        int oldContextLength = 8192;
-        Pair<float[], float[]> ropeFreqs = RoPE.precomputeFreqsCis(config.contextLength, config.headSize, config.ropeTheta,
-                ropeScaling, scaleFactor, loFreqFactor, hiFreqFactor, oldContextLength);
-        float[] ropeFreqsReal = ropeFreqs.first();
-        float[] ropeFreqsImag = ropeFreqs.second();
-
-        GGMLTensorEntry tokenEmbeddings = tensorEntries.get("token_embd.weight");
-        Llama.Weights qw = new Llama.Weights(
-                loadQuantized(tokenEmbeddings),
-                loadArrayOfFloatBuffer(config.numberOfLayers, i -> tensorEntries.get("blk." + i + ".attn_norm.weight")),
-                loadArrayOfQuantized(config.numberOfLayers, i -> tensorEntries.get("blk." + i + ".attn_q.weight")),
-                loadArrayOfQuantized(config.numberOfLayers, i -> tensorEntries.get("blk." + i + ".attn_k.weight")),
-                loadArrayOfQuantized(config.numberOfLayers, i -> tensorEntries.get("blk." + i + ".attn_v.weight")),
-                loadArrayOfQuantized(config.numberOfLayers, i -> tensorEntries.get("blk." + i + ".attn_output.weight")),
-                loadArrayOfFloatBuffer(config.numberOfLayers, i -> tensorEntries.get("blk." + i + ".ffn_norm.weight")),
-                loadArrayOfQuantized(config.numberOfLayers, i -> tensorEntries.get("blk." + i + ".ffn_gate.weight")), // w1
-                loadArrayOfQuantized(config.numberOfLayers, i -> tensorEntries.get("blk." + i + ".ffn_down.weight")), // w2
-                loadArrayOfQuantized(config.numberOfLayers, i -> tensorEntries.get("blk." + i + ".ffn_up.weight")), // w3
-                toFloatBuffer(tensorEntries.get("output_norm.weight")),
-                FloatBuffer.wrap(ropeFreqsReal),
-                FloatBuffer.wrap(ropeFreqsImag),
-                // If "output.weight" is not present then the embedding weights are tied/shared with the decoder.
-                // This is commonly referred as "tie word embeddings".
-                loadQuantized(tensorEntries.getOrDefault("output.weight", tokenEmbeddings))
-        );
-
-        return qw;
-    }
-
-    private static Tokenizer createTokenizer(Map<String, Object> metadata, Vocabulary vocabulary) {
-        String[] mergeLines = (String[]) metadata.get("tokenizer.ggml.merges");
-        List<Pair<Integer, Integer>> merges = Arrays.stream(mergeLines)
-                .map(line -> line.split(" "))
-                .map(parts ->
-                        new Pair<>(
-                                vocabulary.getIndex(parts[0]).orElseThrow(),
-                                vocabulary.getIndex(parts[1]).orElseThrow())
-                ).toList();
-
-        int allTokens = vocabulary.size();
-        int baseTokens = 128000; // assume all tokens after the base ones are special.
-        int reservedSpecialTokens = allTokens - baseTokens;
-        List<String> specialTokensList = Arrays.stream(vocabulary.tokens(), baseTokens, allTokens).toList();
-
-        assert specialTokensList.stream().allMatch(token -> vocabulary.getIndex(token).isPresent());
-
-        Map<String, Integer> specialTokens =
-                IntStream.range(0, specialTokensList.size())
-                        .boxed()
-                        .collect(Collectors.toMap(
-                                i -> specialTokensList.get(i),
-                                i -> baseTokens + i)
-                        );
-
-        return new Tokenizer(vocabulary, merges, LLAMA_3_PATTERN, specialTokens);
-    }
-
-    public static FloatTensor loadQuantized(GGMLTensorEntry entry) {
-        GGMLType ggmlType = entry.ggmlType();
-        return switch (ggmlType) {
-            //case F32 -> new F32FloatTensor(FloatTensor.numberOfElements(entry.shape()), entry.memorySegment());
-            case Q8_0 -> new Q8_0FloatTensor(FloatTensor.numberOfElements(entry.shape()), entry.memorySegment());
-            case Q4_0 -> new Q4_0FloatTensor(FloatTensor.numberOfElements(entry.shape()), entry.memorySegment());
-            case BF16 -> new BF16FloatTensor(FloatTensor.numberOfElements(entry.shape()), entry.memorySegment());
-            case F16 -> new F16FloatTensor(FloatTensor.numberOfElements(entry.shape()), entry.memorySegment());
-            default -> throw new UnsupportedOperationException("Quantization format " + ggmlType);
-        };
-    }
-
-    public static FloatTensor[] loadArrayOfQuantized(int size, IntFunction<GGMLTensorEntry> getTensorEntry) {
-        FloatTensor[] array = new FloatTensor[size];
-        for (int i = 0; i < size; i++) {
-            array[i] = loadQuantized(getTensorEntry.apply(i));
-        }
-        return array;
-    }
-
-    public static FloatBuffer[] loadArrayOfFloatBuffer(int size, IntFunction<GGMLTensorEntry> getTensorEntry) {
-        FloatBuffer[] array = new FloatBuffer[size];
-        for (int i = 0; i < size; i++) {
-            array[i] = toFloatBuffer(getTensorEntry.apply(i));
-        }
-        return array;
-    }
-
-    public static FloatBuffer toFloatBuffer(GGMLTensorEntry tensorEntry) {
-        GGMLType ggmlType = tensorEntry.ggmlType();
-        return switch (ggmlType) {
-            case F32 -> tensorEntry.memorySegment().asByteBuffer().order(ByteOrder.LITTLE_ENDIAN).asFloatBuffer();
-            default -> throw new UnsupportedOperationException("Conversion to " + ggmlType);
-        };
-    }
-}
 
 record Llama(Configuration configuration, Tokenizer tokenizer, Weights weights) {
     public State createNewState(int batchsize) {
@@ -2507,7 +1879,7 @@ final class AOT {
             try (FileChannel fileChannel = FileChannel.open(path, StandardOpenOption.READ)) {
                 return new PartialModel(
                         path.getFileName().toString(),
-                        ModelLoader.loadModel(fileChannel, gguf, Llama3.Options.DEFAULT_MAX_TOKENS, false),
+                        ModelLoader.loadModel(fileChannel, gguf, Options.DEFAULT_MAX_TOKENS, false),
                         gguf.getTensorDataOffset(),
                         gguf.getTensorInfos()
                 );
